+++ cpp/collision/primitive_sdf_binding.cpp
@@ -11,7 +11,7 @@ namespace py = pybind11;
 void bind_primitive_sdf(py::module& m) {
   auto m_psdf = m.def_submodule("primitive_sdf");
   py::class_<Pose>(m_psdf, "Pose", py::module_local())
-      .def(py::init<const Eigen::Vector3d&, const Eigen::Matrix3d&>());
+      .def(py::init<const Eigen::Vector3f&, const Eigen::Matrix3f&>());
   py::class_<AABB>(m_psdf, "AABB")
       .def_readonly("lb", &AABB::lb)
       .def_readonly("ub", &AABB::ub);
@@ -37,28 +37,28 @@ void bind_primitive_sdf(py::module& m) {
       .def("get_aabb", &UnionSDF::get_aabb);
   py::class_<GroundSDF, GroundSDF::Ptr, PrimitiveSDFBase>(m_psdf, "GroundSDF",
                                                           py::module_local())
-      .def(py::init<double>())
+      .def(py::init<float>())
       .def("evaluate_batch", &GroundSDF::evaluate_batch)
       .def("evaluate", &GroundSDF::evaluate)
       .def("is_outside", &GroundSDF::is_outside)
       .def("get_aabb", &GroundSDF::get_aabb);
   py::class_<BoxSDF, BoxSDF::Ptr, ClosedPrimitiveSDFBase>(m_psdf, "BoxSDF",
                                                           py::module_local())
-      .def(py::init<const Eigen::Vector3d&, const Pose&>())
+      .def(py::init<const Eigen::Vector3f&, const Pose&>())
       .def("evaluate_batch", &BoxSDF::evaluate_batch)
       .def("evaluate", &BoxSDF::evaluate)
       .def("is_outside", &BoxSDF::is_outside)
       .def("get_aabb", &BoxSDF::get_aabb);
   py::class_<CylinderSDF, CylinderSDF::Ptr, ClosedPrimitiveSDFBase>(
       m_psdf, "CylinderSDF", py::module_local())
-      .def(py::init<double, double, const Pose&>())
+      .def(py::init<float, float, const Pose&>())
       .def("evaluate_batch", &CylinderSDF::evaluate_batch)
       .def("evaluate", &CylinderSDF::evaluate)
       .def("is_outside", &CylinderSDF::is_outside)
       .def("get_aabb", &CylinderSDF::get_aabb);
   py::class_<SphereSDF, SphereSDF::Ptr, ClosedPrimitiveSDFBase>(
       m_psdf, "SphereSDF", py::module_local())
-      .def(py::init<double, const Pose&>())
+      .def(py::init<float, const Pose&>())
       .def("evaluate_batch", &SphereSDF::evaluate_batch)
       .def("evaluate", &SphereSDF::evaluate)
       .def("is_outside", &SphereSDF::is_outside)
